+++ storage/execout/config.go
@@ -80,3 +80,12 @@ func (c *Config) ListSnapshotFiles(ctx context.Context, inRange *bstream.Range) 
 
 	return files, nil
 }
+
+func (c *Config) ReadFile(ctx context.Context, inrange *block.Range) (*File, error) {
+
+	file := c.NewFile(inrange)
+	if err := file.Load(ctx); err != nil {
+		return nil, err
+	}
+	return file, nil
+}