+++ docs/release-notes/change-log.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Pick up docs from the README.md or README in the same directory as the manifest, when top-level package.doc is empty
 * Tier2 service now supports a maximum concurrent requests limit. Default set to 0 (unlimited). 
 * Improved file listing performance for Google Storage backends by 25%
+* Tier2 will now read back mapper outputs (if they exist) to prevent running them again. Additionally, it will not read back the full blocks if its inputs can be satisfied from existing cached mapper outputs.
+* Tier2 will skip processing completely if the output_module is a mapper that has already been processed (ex: when multiple requests are indexing the same data at the same time)
 
 ## v1.3.7